Proposed rewrite of structure watchers
Solves the TODO about clearing visited locations as well

CC @kenkron

I got bored again. Decided to look at adding compat to some modded
structures and then this happened.
