+++ src/main/java/hunternif/mc/atlas/ext/watcher/DeathWatcher.java
@@ -1,0 +1,26 @@
+package hunternif.mc.atlas.ext.watcher;
+
+import hunternif.mc.atlas.SettingsConfig;
+import hunternif.mc.atlas.api.AtlasAPI;
+import hunternif.mc.atlas.registry.MarkerTypes;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraftforge.event.entity.living.LivingDeathEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+
+/**
+ * Puts an skull marker to the player's death spot.
+ * @author Hunternif
+ */
+public class DeathWatcher {
+	@SubscribeEvent
+	public void onPlayerDeath(LivingDeathEvent event) {
+		if (event.getEntity() instanceof EntityPlayer && SettingsConfig.gameplay.autoDeathMarker) {
+			EntityPlayer player = (EntityPlayer) event.getEntity();
+			for (int atlasID : AtlasAPI.getPlayerAtlases(player)) {
+				AtlasAPI.markers.putMarker(player.getEntityWorld(), true, atlasID, MarkerTypes.TOMB.getRegistryName().toString(),
+						"gui.antiqueatlas.marker.tomb " + player.getName(),
+						(int)player.posX, (int)player.posZ);
+			}
+		}
+	}
+}
